Methods Upload - transitivity reversal
This upload is a jupyter notebook containing methods for determining all of the entries in the matrix that result from closure of the algebraic structure. The input must be a list of lists of integers (zeros or ones), and the output is a list of lists of integers (zeros or ones). Processing of the user input via the gui still needs to be done.